Move all log constants to common component
